Started working on Anypoint styling
