+++ demo/parts.html
@@ -1,0 +1,126 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
+    <title>api-navigation demo</title>
+
+    <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+    <script src="../node_modules/web-animations-js/web-animations-next.min.js"></script>
+    <script type="module" src="../node_modules/@advanced-rest-client/arc-demo-helper/arc-demo-helper.js"></script>
+
+    <style>
+      html,
+      body {
+        margin: 0 !important;
+        padding: 0 !important;
+        font-family: "Open Sans", Helvetica, Arial, sans-serif !important;
+        /* --api-navigation-color: #58595a; */
+        /* --api-navigation-list-item-selected-background-color: #e8e9ea; */
+        /* --api-navigation-list-item-selected-border-left: 4px #00A2DF solid;
+        --api-navigation-list-item-hovered-border-left: 4px #ABE2F5 solid; */
+        --api-navigation-list-item-padding: 5px 15px;
+        /* --api-navigation-list-item-selected-weight: 500; */
+
+        /* --api-navigation-operation-endpoint-opened-background-color: #f4f5f6;
+        --api-navigation-list-item-selected-color: #000;
+        --api-navigation-list-section-font-size: 15px;
+        --api-navigation-list-item-summary-padding: 8px 12px;
+        --api-navigation-endpoint-font-size: 15px; */
+
+        --method-display-selected-color: #fff;
+        /* HTTP methods colors in method documentation panel */
+        --http-method-label-border-radius: 1px;
+        --http-method-label-get-background-color: #00b49d;
+        --http-method-label-get-color: #fff;
+        --http-method-label-post-background-color: #9a63f9;
+        --http-method-label-post-color: #fff;
+        --http-method-label-put-background-color: #F2BE24;
+        --http-method-label-put-color: #fff;
+        --http-method-label-patch-background-color: #5E66F9;
+        --http-method-label-patch-color: #fff;
+        --http-method-label-delete-background-color: #D1344E;
+        --http-method-label-delete-color: #fff;
+        --http-method-label-options-background-color: #00B5D1;
+        --http-method-label-options-color: #fff;
+        --http-method-label-head-background-color: #17BC65;
+        --http-method-label-head-color: #fff;
+        --http-method-label-padding: 4px 5px;
+        --method-display-font-weigth: 100;
+        --dark-divider-opacity: 0;
+      }
+
+      api-navigation::part(api-navigation-list-item) {
+        color: #58595a;
+        border-left: 4px transparent solid;
+        /* padding: 5px 15px; */
+        font-size: 15px;
+      }
+
+      api-navigation::part(api-navigation-list-item):hover {
+        border-left-color: #ABE2F5;
+      }
+
+      api-navigation::part(api-navigation-list-item-selected) {
+        border-left-color: #00A2DF !important;
+        background-color: #e8e9ea;
+        font-weight: 500;
+      }
+
+      /* api-navigation-list-item */
+
+      header {
+        padding: 12px 24px;
+        background-color: #262728;
+        --iron-icon-fill-color: #fff;
+        --paper-input-container-color: #fff;
+        --paper-input-container-input-color: #fff;
+        display: flex;
+        align-items: center;
+      }
+
+      h2 {
+        font-size: 20px;
+        font-weight: 500;
+        line-height: 28px;
+      }
+
+      .box {
+        box-shadow: 0 4px 5px 0 rgba(0, 0, 0, 0.14),
+                    0 1px 10px 0 rgba(0, 0, 0, 0.12),
+                    0 2px 4px -1px rgba(0, 0, 0, 0.4);
+        box-sizing: border-box;
+        padding: 16px;
+        margin-bottom: 40px;
+      }
+
+      .search-container {
+        flex: 1;
+        flex-basis: 0.000000001px;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+      }
+
+      input[type="search"] {
+        font-size: 15px;
+        padding: 8px;
+        min-width: 320px;
+      }
+
+      .search-button {
+        font-size: 15px;
+        padding: 8px;
+      }
+
+      .anypoint-container {
+        background-color: #f9fafb;
+      }
+      </style>
+  </head>
+  <body>
+    <div id="demo"></div>
+
+    <script type="module" src="parts.js"></script>
+  </body>
+</html>
